I added for loops. They work pretty good but tend to bog things down a bit.
I fixed up scope to work a lot likey how python works and now how
I thought it worked. Scope is no longer white-spaced based w.r.t.
variable access. If you define a variable in the script you can use
it in any scope. The scoping for variables is now for temp
vars like for-loops, and for exposing members of a class to the scripts.