Re-write of environment and contents.
Contents is now inventory

in_environment is now simply environment.

All environment attributes return references, eg ch.environment would return the actual instance of the environment.

to_environment and from_environment have been replaced by get and put

To take an item from a room by a character:

room.get(item)
ch.put(item)

each action is internalized to the individual class. Items understand how to add or remove items to themselves, characters and rooms the same.

So you would no longer do, ch.to_environment(room)

Each get/put is over-ridden by the inheriting class.

The get/put is individualized for that class, based on what the class can or cannot do.

The only exception to this rule is that equip and unequip simultaneously pull/place items in the inventory based on whatever action is taking place. So the inventory is the gateway to the equipped dict, to equip a char you would still need to place the item on the character, so:

ch.put(item)
ch.equip(item)